move efile automated messages to state_file namespace
