--- app/models/automated_message/efile_rejected.rb
@@ -1,20 +1,0 @@
-module AutomatedMessage
-  class EfileRejected < AutomatedMessage
-
-    def self.name
-      'messages.efile.rejected'.freeze
-    end
-
-    def sms_body(**args)
-      I18n.t("messages.efile.rejected.sms", **args)
-    end
-
-    def email_subject(**args)
-      I18n.t("messages.efile.rejected.email.subject", **args)
-    end
-
-    def email_body(**args)
-      I18n.t("messages.efile.rejected.email.body", **args)
-    end
-  end
-end
